pyproject.toml
[tool.poetry]
name = "youtube-analytics"
version = "0.1.0"
description = ""
authors = ["Rocksed <ilyagotin@mail.ru>"]
readme = "README.md"
packages = [{include = "youtube_analytics"}]

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^22.2.0"
certifi = "^2022.12.7"
google-api-core = "^2.11.0"
google-api-python-client = "^2.80.0"
google-auth-httplib2 = "^0.1.0"
google-auth = "^2.16.2"
googleapis-common-protos = "^1.58.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
